cockpit-actions: Segregate HTTP Request actions on its own component
For the sake of code maintainability.
